+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - [FEATURE] Add a breadcrumbs as a dashboard title #702
+- [ENHANCEMENT] Dashboard variables list is displayed as a sticky header #703
 
 ## 0.12.0 / 2022-10-25